no-smell: move all vote type counters to Votes collection
Все подсчеты количества голосов разных типов стало возможно перенести
в класс-коллекцию Votes. В результате оказались не нужны многословные
методы, подсчитывавшие голоса в циклах и с помощью полей класса (по
сути, глобальных переменных)

+++ java/src/main/java/org/elections/Fraction.java
@@ -4,7 +4,7 @@ public class Fraction {
     private final long numerator;
     private final long denominator;
 
-    public static Fraction withNumeratorDenominator(long numerator, int denominator) {
+    public static Fraction withNumeratorDenominator(long numerator, long denominator) {
         return new Fraction(numerator, denominator);
     }
